Refactor: Moved tests from the "src" folder to the "test" folder for better project organization.
This commit focuses on improving the structure and organization of the project by moving the tests from the "src" folder to the "test" folder. This change will make it easier for developers to locate and run the tests, as they are now in a dedicated folder.
