+++ package.json
@@ -72,7 +72,7 @@
       "json",
       "ts"
     ],
-    "rootDir": "src",
+    "rootDir": "./",
     "testRegex": ".*\\.spec\\.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
@@ -97,4 +97,5 @@
     },
     "testEnvironment": "node"
   }
+  
 }
