+++ test/application/auth/auth.controller.spec.ts
@@ -1,12 +1,12 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
-import { LoginDto } from './cqrs/dto/login.dto';
-import { RegisterDto } from './cqrs/dto/register.dto';
+import { AuthController } from '../../../src/application/auth/auth.controller';
+import { AuthService } from '../../../src/application/auth/auth.service';
+import { LoginDto } from '../../../src/application/auth/cqrs/dto/login.dto';
+import { RegisterDto } from '../../../src/application/auth/cqrs/dto/register.dto';
 import { mock, MockProxy } from 'jest-mock-extended';
-import { User } from '../../domain/user/user.entity';
-import { GetLoginQuery } from './cqrs/queries';
-import { CreateRegisterCommand } from './cqrs/commands';
+import { User } from '../../../src/domain/user/user.entity';
+import { GetLoginQuery } from '../../../src/application/auth/cqrs/queries';
+import { CreateRegisterCommand } from '../../../src/application/auth/cqrs/commands';
 
 describe('AuthController', () => {
   let controller: AuthController;
